Refactoring Portfolio section
Refactoring Portfolio section

+++ Model/equityHistory.cs
@@ -7,7 +7,7 @@ namespace Git_Sandbox.Model
 	public class equityHistory
 	{
 		public string equityid{ get; set; }
-		public double? price{ get; set; }	
+		public decimal? price{ get; set; }	
 		public int month{ get; set; }
 		public int year { get; set; }		
 		public int assetType { get; set; }
